// backend/src/modules/lesson/interface/dto/lesson-status.enum.ts
export enum LessonInputStatusEnum {
	PENDING_PAID = "PENDING_PAID",
	PENDING_UNPAID = "PENDING_UNPAID",
	COMPLETED_PAID = "COMPLETED_PAID",
	COMPLETED_UNPAID = "COMPLETED_UNPAID",
	RESCHEDULED = "RESCHEDULED",
}

export enum LessonStatusEnum {
	PENDING_UNPAID = "PENDING_UNPAID",
	PENDING_PAID = "PENDING_PAID",
	COMPLETED_UNPAID = "COMPLETED_UNPAID",
	COMPLETED_PAID = "COMPLETED_PAID",
	MISSED = "MISSED",
	RESCHEDULED = "RESCHEDULED",
	CANCELLED = "CANCELLED",
}